SL-1136: fix tests, remove unused service
